(refactor) Move tooltip helper functions to new file
